nexu_test/requirements.txt
django==3.0.6
djangorestframework==3.11.0
psycopg2-binary==2.8.5
